requirements/development.txt
-r base.txt

black==20.8b1
coverage==5.3
django-stubs==1.6.0
flake8==3.8.4
isort==5.5.4
mypy==0.782
